+++ crates/chain/Cargo.toml
@@ -14,10 +14,8 @@ readme = "README.md"
 
 [dependencies]
 bitcoin = { version = "0.32.0", default-features = false }
+bdk_core = { path = "../core", version = "0.1", default-features = false }
 serde = { version = "1", optional = true, features = ["derive", "rc"] }
-
-# Use hashbrown as a feature flag to have HashSet and HashMap from it.
-hashbrown = { version = "0.9.1", optional = true, features = ["serde"] }
 miniscript = { version = "12.0.0", optional = true, default-features = false }
 
 # Feature dependencies
@@ -30,6 +28,7 @@ proptest = "1.2.0"
 
 [features]
 default = ["std", "miniscript"]
-std = ["bitcoin/std", "miniscript?/std"]
-serde = ["dep:serde", "bitcoin/serde", "miniscript?/serde"]
+std = ["bitcoin/std", "miniscript?/std", "bdk_core/std"]
+serde = ["dep:serde", "bitcoin/serde", "miniscript?/serde", "bdk_core/serde"]
+hashbrown = ["bdk_core/hashbrown"]
 rusqlite = ["std", "dep:rusqlite", "serde", "serde_json"]
