+++ crates/core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "bdk_core"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.63"
+homepage = "https://bitcoindevkit.org"
+repository = "https://github.com/bitcoindevkit/bdk"
+documentation = "https://docs.rs/bdk_core"
+description = "Collection of core structures for Bitcoin Dev Kit."
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+
+[dependencies]
+bitcoin = { version = "0.32", default-features = false }
+serde = { version = "1", optional = true, features = ["derive", "rc"] }
+hashbrown = { version = "0.9.1", optional = true }
+
+[features]
+default = ["std"]
+std = ["bitcoin/std"]
+serde = ["dep:serde", "bitcoin/serde", "hashbrown?/serde"]
